Fix PR review feedback for dogfood cleanup
